feat: [NVS-208] 로그인 회원가입 ui 변경
